Adjust menu and initial settings page
